Extract I2_S QK256 forward dispatch into SRP microcrate
