+++ crates/bitnet-qk256-dispatch/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "bitnet-qk256-dispatch"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "SRP microcrate for I2_S QK256 forward dispatch over Candle tensors"
+categories.workspace = true
+keywords.workspace = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-quantization = { path = "../bitnet-quantization", version = "0.2.1-dev" }
+candle-core = { workspace = true }
+
+[dev-dependencies]
+bitnet-test-support.workspace = true
+
+[features]
+default = []
+cpu = ["bitnet-quantization/cpu"]
+cuda = ["bitnet-common/cuda", "candle-core/cuda", "bitnet-quantization/cuda"]
+metal = ["bitnet-common/metal"]
+gpu = ["cuda", "metal"]
