+++ crates/bitnet-transformer/Cargo.toml
@@ -14,6 +14,7 @@ rust-version.workspace = true
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
 bitnet-quantization = { path = "../bitnet-quantization", version = "0.2.1-dev" }
+bitnet-qk256-dispatch = { path = "../bitnet-qk256-dispatch", version = "0.2.1-dev" }
 bitnet-rope = { path = "../bitnet-rope", version = "0.2.1-dev" }
 candle-core = { workspace = true }
 candle-nn = { workspace = true }
@@ -29,8 +30,8 @@ insta.workspace = true
 
 [features]
 default = []
-cpu = ["bitnet-quantization/cpu"]
-cuda = ["bitnet-common/cuda", "candle-core/cuda", "bitnet-quantization/cuda"]
-metal = ["bitnet-common/metal"]
+cpu = ["bitnet-quantization/cpu", "bitnet-qk256-dispatch/cpu"]
+cuda = ["bitnet-common/cuda", "candle-core/cuda", "bitnet-quantization/cuda", "bitnet-qk256-dispatch/cuda"]
+metal = ["bitnet-common/metal", "bitnet-qk256-dispatch/metal"]
 gpu = ["cuda", "metal"]
 trace = []
