+++ crates/bitnet-vulkan/src/lib.rs
@@ -6,8 +6,6 @@
 pub use bitnet_vulkan_shaders::VulkanShaderSource;
 
 /// Backward-compatible kernel module re-export.
-pub mod kernels {
-    pub use bitnet_vulkan_shaders::VulkanShaderSource;
-}
+pub mod kernels {}
 
 pub mod runtime;
